% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface_elements.R
\name{box_interfejs1}
\alias{box_interfejs1}
\title{Create tile for the first interface}
\usage{
box_interfejs1(id, text)
}
\arguments{
\item{id}{`string` - id boxa w CSS}

\item{text}{`string` - Teks w środku boxa}
}
\value{
box z tekstem
}
\description{
Używany w app_ui.R. Generuje boxy widoczne na pierwszym interfejsie aplikacji.
}
