% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{create_filters_elements}
\alias{create_filters_elements}
\title{Create filters as sliders and drop-down lists}
\usage{
create_filters_elements(attr_list)
}
\value{
named list of filters
}
\description{
Create filters as sliders and drop-down lists
}
