% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dokumentacja.R
\docType{data}
\name{Electricity}
\alias{Electricity}
\title{Informacje o zużyciu prądu}
\format{
An object of class \code{data.frame} with 1051200 rows and 25 columns.
}
\usage{
Electricity
}
\description{
Ramka zawiera informacje o zużywanym prądzie przez urządzenia użytkownika w jego gospodarstwie domowym.
}
\references{
\url{TODO}
}
\author{
Piotr Sieńko
}
\keyword{data}
