% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{filter_by_attr}
\alias{filter_by_attr}
\title{Filter given dataset using a list of filters ('numeric' or 'factor' type)}
\usage{
filter_by_attr(filters, dataset)
}
\value{
filtered dataset
}
\description{
Filter given dataset using a list of filters ('numeric' or 'factor' type)
}
\details{
it is assumed that filters variable has the same structure as in output of get_attr_info function
A named `list`. Each element describes 1 filter as a `list` with 3 elements:
 * `name` - a single string with name of attribute to filter by
 * `type` - possible values: "numeric" or "factor"
 * `range` - for "numeric": c(min, max); for "factor" - levels vector to filter by
}
