% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_attr_info}
\alias{get_attr_info}
\title{Get info about attributes}
\usage{
get_attr_info(dataset)
}
\arguments{
\item{dataset}{A single `string`, equal to one of the names of datasets. 
For now, only 'fridges' would be accepted OR
A `data.frame`, following format desribed in Details.}
}
\value{
A named `list`. Each element describes 1 attribute as a `list` with 3 elements:
 * `name` - a single string with name of attribute
 * `type` - possible values: "numeric" or "factor"
 * `range` - for "numeric": c(min, max); for "factor" - levels vector
}
\description{
Get info about attributes
}
\details{
It is assumed, that each dataset is 
a `data.frame`, where ALL columns following column `Zdj` describe attributes.
Naturally, it has to contain EXACTLY ONE column `Zdj`.
}
