% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_best_fridges}
\alias{get_best_fridges}
\title{Get best fridges}
\usage{
get_best_fridges(cur_m_power, el_cost, top_n = 5, filters = NA)
}
\value{
A data.frame with 4 columns: ID, Nazwa, Cena, Roczne_zuzycie_pradu_kWh
with info about most cost-efficient top_n fridges
}
\description{
Get best fridges
}
