% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_best_models}
\alias{get_best_models}
\title{Najlepsze urzadzenia}
\usage{
get_best_models(
  urzadzenie,
  cur_m_power,
  el_cost,
  tv_con,
  top_n = 5,
  filters = NA,
  criterion = "years_to_go"
)
}
\arguments{
\item{urzadzenie}{typ urzadzenia do wyswietlenia}

\item{cur_m_power}{moc aktualne posiadanego przez urzytkownika uzadzenia}

\item{el_cost}{koszt pradu, ktore zuzywaja uzadzenia}

\item{top_n}{liczba najleprzych modeli do pokazania}

\item{filters}{filtry do zaaplikowania (być może NA)}

\item{criterion}{String, jeden z 3: "years_to_go" (domyślnie), "power_efficiency", "prize".
years_to_go - ile czasu czekamy do zwrotu inwestycji
power_efficiency - co zuzyje najmniej pradu wg prognozy
prize - co najmniej kosztuje sposrod tych, ktore sa bardziej energooszczedne od aktualnego}
}
\value{
Patrz `get_best_fridges` oraz `get_best_tvs`
}
\description{
Znajdz i zwroc najlepsze urzadzenia w zaleznosci od parametru `criterion`.
}
