% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_fridge_con}
\alias{get_fridge_con}
\title{Calculate power consumption per month}
\usage{
get_fridge_con()
}
\value{
monthly energy consumption data.frame with: 
Month - month number as numeric
FGE - monthly average power level in W as numeric
kWh - monthly energy usage in kWh as numeric
}
\description{
Calculate power consumption per month
}
