% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_new_tv_con}
\alias{get_new_tv_con}
\title{Calculate power consumption per month for new tv}
\usage{
get_new_tv_con(stand_con, on_con, monthly_con)
}
\arguments{
\item{stand_con}{stanby power consumption}

\item{on_con}{running power consumption}

\item{monthly_con}{dataset produced by get_fridge_con}
}
\value{
monthly energy consumption vector in kWh
}
\description{
Calculate power consumption per month for new tv
}
