% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_offers}
\alias{get_offers}
\title{Get offers for a given model and type of device}
\usage{
get_offers(id, type_of_device)
}
\arguments{
\item{id}{id of the device for which we want to get offers}

\item{type_of_device}{`string` that determines the type of device.
For now, only 'fridges' or 'tvs' would be accepted}
}
\value{
`list`` with four elements, every of which contains: Cena, Sklep, URL, NR
}
\description{
Get offers for a given model and type of device
}
