% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_tv_con}
\alias{get_tv_con}
\title{Calculate power consumption and usage time per month for current tv}
\usage{
get_tv_con()
}
\value{
monthly energy consumption data.frame with: 
Month - month number as numeric
TVE - monthly average power level in W as numeric
kWh - monthly energy usage in kWh as numeric 
Off - average number of minutes with turned off tv /each month/ as 
On - average monthly number of minutes with turned on tv /each month
}
\description{
Calculate power consumption and usage time per month for current tv
}
