% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{kolejnosc_kryteriow}
\alias{kolejnosc_kryteriow}
\title{Ustala kolejnosc kryterium od najleprzego do najgorszego}
\usage{
kolejnosc_kryteriow(wartosci_kryteriow, criterion)
}
\arguments{
\item{wartosci_kryteriow}{Wartosci zwracane przez `get_best_models` w kolumnie `criterion`}

\item{criterion}{String, jeden z 4: "years_to_go" (domyślnie), "power_efficiency", "prize", "true_cost".
years_to_go - ile czasu czekamy do zwrotu inwestycji
power_efficiency - największa różnica w MIESIĘCZNYM zużyciu prądu
prize - co najmniej kosztuje sposrod tych, ktore sa bardziej energooszczedne od aktualnego
true_cost - cena produktu minus zysk na kosztach pradu w poruwnaniu z aktualnym modelem}
}
\value{
permuteację zbioru 1:n ustalajacego kolenosc `wartosci_kryteriow`
}
\description{
Wedlog jednego z kryteriow leprze sa te modele, ktore maja wartosc najwieksza, wedlog pozostalych dwoch - najmniejsza
}
