% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{list_el}
\alias{list_el}
\title{Create drop-down list}
\usage{
list_el(name, range)
}
\arguments{
\item{name}{`string` - nazwa tworzonego slidera}

\item{range}{wektor `factor` zawierajacy mozliwe do wyboru opcje}
}
\value{
selectInput from shiny package based on given label (parametr name) and choices (parametr range)
}
\description{
Create drop-down list
}
