% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{slider_el}
\alias{slider_el}
\title{Create slider input}
\usage{
slider_el(name, range)
}
\value{
sliderInput from shiny package based on given label (name) and range
}
\description{
Create slider input
}
