% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{tekst_do_wyswietlania}
\alias{tekst_do_wyswietlania}
\title{Ustala jakie napisy powinny sie wyswietlac na kafelkach interfejsu 1}
\usage{
tekst_do_wyswietlania(wartosci_kryteriow, criterion)
}
\arguments{
\item{wartosci_kryteriow}{Wartosci zwracane przez `get_best_models` w kolumnie `criterion`}

\item{criterion}{String, jeden z 3: "years_to_go" (domyślnie), "power_efficiency", "prize".
years_to_go - ile czasu czekamy do zwrotu inwestycji
power_efficiency - największa różnica w MIESIĘCZNYM zużyciu prądu
prize - co najmniej kosztuje sposrod tych, ktore sa bardziej energooszczedne od aktualnego}
}
\value{
wektor napisow majacych byc wyswietlane na interfejsie 1
}
\description{
Ustala jakie napisy powinny sie wyswietlac na kafelkach interfejsu 1
}
