% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dokumentacja.R
\docType{data}
\name{tvs}
\alias{tvs}
\title{Energooszczędne lodówki}
\format{
An object of class \code{data.frame} with 49 rows and 20 columns.
}
\usage{
tvs
}
\description{
Ramka zawiera zbiór telewizorów energooszczędnych możliwych do zaproponowania użytkownikowi.
}
\references{
\url{TODO}
}
\author{
Paweł Morgen
}
\keyword{data}
