% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prices.R
\name{update_prices}
\alias{update_prices}
\title{Updates prices to the minimum of the proposed offers for a given type of device}
\usage{
update_prices(type_of_device)
}
\arguments{
\item{type_of_device}{`string` that determines the type of device.
For now, only 'fridges' or 'tvs' would be accepted}
}
\value{

}
\description{
Updates prices to the minimum of the proposed offers for a given type of device
}
