% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{wyswietl_date}
\alias{wyswietl_date}
\title{Zamienia date wyrazona w latach (np. 2.5 lat) na liczbe lat i miesiecy (np. 2 lata i 6 miesiecy)}
\usage{
wyswietl_date(num_years)
}
\arguments{
\item{num_years}{liczba lat}
}
\value{
tekst z odpowiednia liczba lat i miesiecy
}
\description{
Zamienia date wyrazona w latach (np. 2.5 lat) na liczbe lat i miesiecy (np. 2 lata i 6 miesiecy)
}
