% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{yearly_data_to_plot}
\alias{yearly_data_to_plot}
\title{yearly_data_to_plot}
\usage{
yearly_data_to_plot(
  cur_m_power,
  new_m_power,
  new_m_price,
  el_cost,
  cur_month_power,
  new_month_power
)
}
\arguments{
\item{cur_m_power}{roczne zużycie prądu aktualnie posiadanego przez urzytkownika urzadzenia}

\item{new_m_power}{roczne zużycie prądu proponowanego przez urzytkownika urzadzenia}

\item{new_m_price}{cena proponowanego urzadzenia}

\item{el_cost}{koszt pradu, ktore zuzywaja urzadzenia}

\item{cur_month_power}{tabela zuzycia pradu przez aktualne urzadzenie w zaleznosci od miesiaca}

\item{new_month_powera}{wektor zuzycia pradu przez aktualne urzadzenie w zaleznosci od miesiaca}
}
\value{
dataframe
}
\description{
Tworzy dataset potrzebny do wykresu TV
}
\examples{
stand_con <- 0.5
on_con <- 200
cur_month_power <- get_tv_con() - from server_utils
new_month_power <- get_new_tv_con(stand_con, on_con, cur_month_power) - from server_utils
new_m_power <- sum(new_month_power)
cur_m_power <- sum(cur_month_power$kWh)
el_cost <- 0.617
new_m_price <- 1000
yearly_data_to_plot(cur_m_power, new_m_power, new_m_price, el_cost, cur_month_power, new_month_power)
}
