% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{yearly_forecast_plot}
\alias{yearly_forecast_plot}
\title{Yearly forecast plot}
\usage{
yearly_forecast_plot(
  cur_m_power,
  new_m_power,
  new_m_price,
  el_cost,
  cur_month_power
)
}
\description{
Yearly forecast plot
}
