% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dokumentacja.R
\docType{data}
\name{fridges}
\alias{fridges}
\title{Energooszczędne lodówki}
\format{
An object of class \code{data.frame} with 50 rows and 21 columns.
}
\usage{
fridges
}
\description{
Ramka zawiera zbiór logówek energooszczędnych możliwych do zaproponowania użytkownikowi. Zawiera wszystkie ich paametry oraz link do sklepu.
}
\references{
\url{TODO}
}
\author{
Paweł Morgen
}
\keyword{data}
