% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_best_fridges}
\alias{get_best_fridges}
\title{Najlepsze lodowki}
\usage{
get_best_fridges(cur_m_power, el_cost, top_n = 5, filters = NA, criterion)
}
\arguments{
\item{cur_m_power}{roczne zuzycie pradu przez urzadzenie}

\item{el_cost}{koszt 1kWh pradu w domostwie}

\item{top_n}{liczba najleprzych modeli do pokazania}

\item{filters}{filtry do zaaplikowania (być może NA)}

\item{criterion}{String, jeden z 4: "years_to_go" (domyślnie), "power_efficiency", "prize", "true_cost".
years_to_go - ile czasu czekamy do zwrotu inwestycji
power_efficiency - największa różnica w MIESIĘCZNYM zużyciu prądu
prize - co najmniej kosztuje sposrod tych, ktore sa bardziej energooszczedne od aktualnego
true_cost - cena produktu minus zysk na kosztach pradu w poruwnaniu z aktualnym modelem}
}
\value{
`data.frame` z kolumnami: ID, Nazwa, Cena, Roczne_zuzycie_pradu_kWh, criterion
z informacjami o najlepszych lodowkach wg parametru `criterion`.
}
\description{
Znajdz najlepsze lodowki wedlug `criterion`.
}
