% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_best_models}
\alias{get_best_models}
\title{Get best models}
\usage{
get_best_models(
  urzadzenie,
  cur_m_power,
  el_cost,
  tv_con,
  top_n = 5,
  filters = NA
)
}
\arguments{
\item{urzadzenie}{typ urzadzenia do wyswietlenia}

\item{cur_m_power}{moc aktualne posiadanego przez urzytkownika uzadzenia}

\item{el_cost}{koszt pradu, ktore zuzywaja uzadzenia}

\item{top_n}{liczba najleprzych modeli do pokazania}

\item{filters}{filtry do zaaplikowania (być może NA)}
}
\value{
A data.frame with 4 columns: ID, Nazwa, Cena, Roczne_zuzycie_pradu_kWh
with info about most cost-efficient top_n fridges
}
\description{
Get best models
}
