% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_best_models}
\alias{get_best_models}
\title{Najlepsze urzadzenia}
\usage{
get_best_models(
  urzadzenie,
  cur_m_power_fridge,
  tv_con,
  el_cost,
  top_n = 5,
  filters = NA,
  criterion = "years_to_go"
)
}
\arguments{
\item{urzadzenie}{typ urzadzenia do wyswietlenia}

\item{el_cost}{koszt 1kWh pradu w domostwie}

\item{top_n}{liczba najleprzych modeli do pokazania}

\item{filters}{filtry do zaaplikowania (być może NA)}

\item{criterion}{String, jeden z 4: "years_to_go" (domyślnie), "power_efficiency", "prize", "true_cost".
years_to_go - ile czasu czekamy do zwrotu inwestycji
power_efficiency - największa różnica w MIESIĘCZNYM zużyciu prądu
prize - co najmniej kosztuje sposrod tych, ktore sa bardziej energooszczedne od aktualnego
true_cost - cena produktu minus zysk na kosztach pradu w poruwnaniu z aktualnym modelem}

\item{cur_m_power}{roczne zuzycie pradu przez urzadzenie}
}
\value{
Patrz `get_best_fridges` oraz `get_best_tvs`
}
\description{
Znajdz i zwroc najlepsze urzadzenia w zaleznosci od parametru `criterion`.
}
