% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_best_tvs}
\alias{get_best_tvs}
\title{Najlepsze telewizory}
\usage{
get_best_tvs(cur_m_power, el_cost, tv_con, top_n = 5, filters = NA, criterion)
}
\arguments{
\item{cur_m_power}{moc aktualne posiadanego przez urzytkownika uzadzenia}

\item{el_cost}{koszt pradu, ktore zuzywaja uzadzenia}

\item{top_n}{liczba najleprzych modeli do pokazania}

\item{filters}{filtry do zaaplikowania (być może NA)}

\item{criterion}{String, jeden z 3: "years_to_go" (domyślnie), "power_efficiency", "prize".
years_to_go - ile czasu czekamy do zwrotu inwestycji
power_efficiency - co zuzyje najmniej pradu wg prognozy
prize - co najmniej kosztuje sposrod tych, ktore sa bardziej energooszczedne od aktualnego}
}
\value{
`data.frame` z kolumnami: ID, Nazwa, Cena, Roczne_zuzycie_pradu_kWh,Pobor_mocy_tryb_czuwania_W, Pobor_mocy_tryb_wlaczenia_W, years_to_go, criterion
z informacjami o najlepszych telewizorach wg parametru `criterion`.
}
\description{
Znajdz najlepsze telewizory wedlug `criterion`.
}
