% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_kolejnosc_kafelkow_interface1}
\alias{get_kolejnosc_kafelkow_interface1}
\title{Ustala kolejnosc kafelkow na interfejsie 1}
\usage{
get_kolejnosc_kafelkow_interface1(
  cur_m_power_fridge,
  tv_con,
  el_cost,
  criterion
)
}
\arguments{
\item{el_cost}{koszt 1kWh pradu w domostwie}

\item{criterion}{String, jeden z 3: "years_to_go" (domyślnie), "power_efficiency", "prize".
years_to_go - ile czasu czekamy do zwrotu inwestycji
power_efficiency - co zuzyje najmniej pradu wg prognozy
prize - co najmniej kosztuje sposrod tych, ktore sa bardziej energooszczedne od aktualnego}

\item{cur_m_powers}{wektor aktualnego rocznego zurzycia pradu przez kolejne urzadzenia}
}
\value{
wektor `c("fridges", "tvs", "kettles", "washing-machines", "air-conditionings", "microwaves")` posortowanych pod wzgledem danego kryterium.
}
\description{
Na podstawie aktulanego zurzycia energii znajduje najkorzystniejsze wymiany ze wzgledu na wybrane przez uzytkownika kryterium
}
