% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{get_years_to_go}
\alias{get_years_to_go}
\alias{get_true_cost}
\title{Policz opłacalność}
\usage{
get_years_to_go(cur_m_power, new_m_power, new_m_price, el_cost)

get_true_cost(cur_m_power, new_m_power, new_m_price, el_cost)
}
\arguments{
\item{cur_m_power}{moc aktualne posiadanego przez urzytkownika uzadzenia}

\item{new_m_power}{moc proponowanego uzadzenia}

\item{new_m_price}{cena proponowanego uzadzenia}

\item{el_cost}{koszt pradu, ktore zuzywaja uzadzenia}
}
\value{
ilość lat, po których zakup się zwróci (`get_years_to_go`) lub bilans po 5 latach (`get_true_cost`)
}
\description{
Policz opłacalność
}
\section{Functions}{
\itemize{
\item \code{get_years_to_go}: Policz czas do zwrotu inwestycji

\item \code{get_true_cost}: Bilans po pięciu latach
}}

\examples{
cur_m_power <- 400 # kWh/rok
new_m_power <- 170 # kWh/rok
new_m_price <- 1200 # PLN
el_cost <- 0.617 # PLN/kWh, średni koszt energii elektrycznej w Polsce (maj 2020)
test_plot <- yearly_forecast_plot(cur_m_power, new_m_power, new_m_price, el_cost, get_fridge_con())
show(test_plot)
}
