% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{get_years_to_go}
\alias{get_years_to_go}
\title{Calculate cost-effectiveness}
\usage{
get_years_to_go(cur_m_power, new_m_power, new_m_price, el_cost)
}
\description{
Return, in how many years the investment will pay off
}
