% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wykres.R
\name{yearly_forecast_plot}
\alias{yearly_forecast_plot}
\title{Yearly forecast plot}
\usage{
yearly_forecast_plot(
  cur_m_power,
  new_m_power,
  new_m_price,
  el_cost,
  cur_month_power
)
}
\arguments{
\item{cur_m_power}{moc aktualne posiadanego przez urzytkownika uzadzenia}

\item{new_m_power}{moc proponowanego uzadzenia}

\item{new_m_price}{cena proponowanego uzadzenia}

\item{el_cost}{koszt pradu, ktore zuzywaja uzadzenia}

\item{cur_month_power}{tabela zurzycia pradu przez aktualne uzadzenie w zaleznosci od miesiaca}
}
\value{
wykres w pplot2
}
\description{
Tworzy wykres predykcji
}
\examples{
yearly_forecast_plot(400, 170, 1200, 0.617, get_fridge_con())
}
